requirements.txt
# step 1: Install Flask
flask
# step 3: Install Flask-SQLAlchemy
flask-sqlalchemy
# step 11: i18nitialize Flask-Babel
flask-babel==4.0.0
deep_translator
# step 14 - 05: Install Flask-Migrate
polib
# step 14 - llol_step06:
python-dotenv
psycopg2-binary


